tableau_chatbot/requirements.txt
openai
langchain
langchain-openai
langchain-community
faiss-cpu
-e .
nemoguardrails